# api-spec.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Heroku-Salesforce API
  description: Example Heroku app as an API using Salesforce SDK.
servers:
  - url: http://127.0.0.1:3000
paths:
  /generatePDF:
    post:
      operationId: UnitOfWork
      description:
        Receives a payload containing Account, Contact, and Case details and uses the
        Unit of Work pattern to assign the corresponding values to to its Record
        while maintaining the relationships. It then commits the unit of work and
        returns the Record Id's for each object.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                objectId:
                  type: string
              properties:
                accountName:
                  type: string
                lastName:
                  type: string
                subject:
                  type: string
                callbackUrl:
                  type: string
      responses:
        '201':
          description: Received UnitOfWork API request
        '500':
          $ref: '#/components/responses/DeveloperError'
        '503':
          $ref: '#/components/responses/UnknownError'
      callbacks:
        unitOfWorkResponse:
          '{$request.body#/callbackUrl}':
            post:
              description: Response from /unitofwork API
              operationId: generatePDFResponseHandler
              requestBody:
                content:
                  application/json:
                    schema:
                      type: object
                      properties:
                        accountId:
                          type: string
                        contactId:
                          type: string
                        cases:
                          type: object
                          properties:
                            serviceCaseId:
                              type: string
                            followupCaseId:
                              type: string
              responses:   # Expected responses to the callback message
                '200':
                  description: Your server returns this code if it accepts the callback
components:
  responses:
    DeveloperError:
      description: DeveloperError
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnknownError:
      description: UnknownError
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    # Schema for error response body
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message